+++ src/http.c
@@ -678,7 +678,7 @@ httpSendComplete(int fd, char *bufnotused, size_t size, int errflag, void *data)
     StoreEntry *entry = httpState->entry;
     ErrorState *err;
     debug(11, 5) ("httpSendComplete: FD %d: size %d: errflag %d.\n",
-	fd, size, errflag);
+	fd, (int) size, errflag);
 #if URL_CHECKSUM_DEBUG
     assert(entry->mem_obj->chksum == url_checksum(entry->mem_obj->url));
 #endif
@@ -1103,7 +1103,7 @@ httpSendRequestEntry(int fd, char *bufnotused, size_t size, int errflag, void *d
     StoreEntry *entry = httpState->entry;
     ErrorState *err;
     debug(11, 5) ("httpSendRequestEntry: FD %d: size %d: errflag %d.\n",
-	fd, size, errflag);
+	fd, (int) size, errflag);
     if (size > 0) {
 	fd_bytes(fd, size, FD_WRITE);
 	kb_incr(&statCounter.server.all.kbytes_out, size);
