+++ src/store_dir.c
@@ -290,8 +290,8 @@ storeDirStats(StoreEntry * sentry)
     storeAppendPrintf(sentry, "Store Directory Statistics:\n");
     storeAppendPrintf(sentry, "Store Entries          : %d\n",
 	memInUse(MEM_STOREENTRY));
-    storeAppendPrintf(sentry, "Maximum Swap Size      : %8d KB\n",
-	Config.Swap.maxSize);
+    storeAppendPrintf(sentry, "Maximum Swap Size      : %8ld KB\n",
+	(long int) Config.Swap.maxSize);
     storeAppendPrintf(sentry, "Current Store Swap Size: %8d KB\n",
 	store_swap_size);
     storeAppendPrintf(sentry, "Current Capacity       : %d%% used, %d%% free\n",
